feat: update EnumExpr to use i32 for Const variant and enhance InfRational visibility

+++ src/smt/ast.rs
@@ -92,7 +92,7 @@ impl fmt::Display for BoolExpr {
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub enum EnumExpr {
     Var(usize),
-    Const(usize),
+    Const(i32),
 }
 
 impl fmt::Display for EnumExpr {
